feat: Add HTTP request to generate lecture materials
The code changes in `record_lecture.py` include importing the `requests` module and adding an HTTP request to generate lecture materials. This request is sent to `http://localhost:7071/api/ContentPipeline` with the parameters `lecture=celbio_3` and `run_full_pipeline=true`. The request is made within the `show_spinner_till_generated` method, which is called to generate materials.